Add FocusedTextBox abstraction.
