+++ osu.Game/Screens/Select/SearchTextBox.cs
@@ -16,26 +16,8 @@ namespace osu.Game.Screens.Select
     /// <summary>
     /// A textbox which holds focus eagerly.
     /// </summary>
-    public class SearchTextBox : OsuTextBox
+    public class SearchTextBox : FocusedTextBox
     {
-        protected override Color4 BackgroundUnfocused => new Color4(10, 10, 10, 255);
-        protected override Color4 BackgroundFocused => new Color4(10, 10, 10, 255);
-        public Action Exit;
-
-        private bool focus;
-        public bool HoldFocus
-        {
-            get { return focus; }
-            set
-            {
-                focus = value;
-                if (!focus)
-                    TriggerFocusLost();
-            }
-        }
-
-        public override bool RequestingFocus => HoldFocus;
-
         public SearchTextBox()
         {
             Height = 35;
@@ -51,25 +33,6 @@ namespace osu.Game.Screens.Select
             });
 
             PlaceholderText = "type to search";
-        }
-
-        protected override bool OnFocus(InputState state)
-        {
-            var result = base.OnFocus(state);
-            BorderThickness = 0;
-            return result;
-        }
-
-        protected override void OnFocusLost(InputState state)
-        {
-            if (state.Keyboard.Keys.Any(key => key == Key.Escape))
-            {
-                if (Text.Length > 0)
-                    Text = string.Empty;
-                else
-                    Exit?.Invoke();
-            }
-            base.OnFocusLost(state);
         }
 
         protected override bool OnKeyDown(InputState state, KeyDownEventArgs args)
